Fix star styling and placement

+++ brainstorm_tools/models.py
@@ -25,7 +25,7 @@ class Project(models.Model):
 class ProjectRating(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     project = models.ForeignKey(Project, on_delete=models.CASCADE)
-    rating = models.FloatField(default=0.0)
+    rating = models.IntegerField(default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
